refactor: 컨트롤러에서 게임을 진행하는 부분 분리(리뷰 내용 반영)
https://github.com/next-step/java-bowling/pull/544#discussion_r625704740
