+++ docs/bowling.md
@@ -216,6 +216,4 @@
 - step4
     - [x] for문에서 i로 파라미터를 받는 부분 수정
     - [x] `Pins`의 이름을 `BallThrows`로 변경
-    - [ ] 메서드의 깊이가 너무 길어지지 않도록 구현해보기
-        - 뎁스 1단계를 지키는 형태에서는 개선이 불가능한 느낌이다.
-    - [ ] 컨트롤러에서 게임을 진행하는 부분을 분리해보기
+    - [x] 컨트롤러에서 게임을 진행하는 부분을 분리해보기
