+++ src/main/java/bowling/controller/BowlingController.java
@@ -1,12 +1,10 @@
 package bowling.controller;
 
-import bowling.domain.frame.RoundNumber;
-import bowling.domain.pin.Pin;
 import bowling.domain.player.Player;
 import bowling.domain.player.PlayerName;
 import bowling.domain.player.Players;
-import bowling.exception.BowlingException;
 import bowling.exception.PlayerNameValidationException;
+import bowling.service.BowlingGame;
 import bowling.view.InputView;
 import bowling.view.OutputView;
 
@@ -25,10 +23,9 @@ public final class BowlingController {
     }
 
     public void run() {
-        final Players players = players();
-        outputView.printScoreBoard(players);
-
-        playGame(players);
+        final BowlingGame bowlingGame = new BowlingGame(players(), inputView, outputView);
+        bowlingGame.printScoreBoard();
+        bowlingGame.playGame();
     }
 
     private Players players() {
@@ -40,42 +37,12 @@ public final class BowlingController {
         return players;
     }
 
-    private void playGame(Players players) {
-        for (int i = RoundNumber.MIN; i <= RoundNumber.MAX; i++) {
-            bowlAll(players, i);
-        }
-    }
-
-    private void bowlAll(Players players, int roundNumber) {
-        for (Player player : players.value()) {
-            bowl(players, player, roundNumber);
-        }
-    }
-
-    private void bowl(Players players, Player player, int roundNumberSource) {
-        final RoundNumber roundNumber = new RoundNumber(roundNumberSource);
-        while (!player.isEnded(roundNumber)) {
-            downPin(player, roundNumber);
-            outputView.printScoreBoard(players);
-        }
-    }
-
     private PlayerName playerName(int playerCount) {
         try {
             return PlayerName.valueOf(inputView.inputPlayerName(playerCount));
         } catch (PlayerNameValidationException e) {
             System.err.println(e.getMessage());
             return playerName(playerCount);
-        }
-    }
-
-    private void downPin(final Player player, final RoundNumber roundNumber) {
-        try {
-            final Pin pin = new Pin(inputView.inputDownPin(player));
-            player.knockDownPin(roundNumber, pin);
-        } catch (BowlingException e) {
-            System.err.println(e.getMessage());
-            downPin(player, roundNumber);
         }
     }
 }
